backend/requirements.txt
Flask
pyodbc
python-dotenv
flask_cors
flask-login